fix-controller-class: user controller class change directory
